+++ src/utilities/director-assistant.ts
